Directory rearranging. Amended User class. Created Account class, adjusted ViewModel's HashMap value type
