+++ app/src/main/java/com/example/coin_panion/HistoryFragment.java
@@ -10,10 +10,9 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
-import com.example.coin_panion.classes.Notification;
-import com.example.coin_panion.classes.NotificationAdapter;
+import com.example.coin_panion.classes.notification.Notification;
+import com.example.coin_panion.classes.notification.NotificationAdapter;
 
-import java.lang.reflect.Array;
 import java.util.ArrayList;
 import java.util.List;
