+++ app/src/main/java/com/example/coin_panion/classes/friends/Friend.java
@@ -1,77 +1,68 @@
 package com.example.coin_panion.classes.friends;
 
-import com.example.coin_panion.classes.Notification;
-import com.example.coin_panion.classes.settleUp.SettleUpDebt;
-
 import java.sql.Blob;
+import java.util.List;
 
 public class Friend{
 
+    /*Friend Info*/
     private boolean blockedStatus;
     private Integer friendID;
-    private Integer userID;
+    private Integer accountID;
     private String friendName;
     private String phoneNumber;
     private Blob profilePic;
 
-    public Friend(Integer userID, Integer friendID, String friendName, String phoneNumber, Blob profilePic) {
+    /*Additional Friend info*/
+    //TODO how to know if that friend is in debt with us
+
+
+    /*Create new friend object based on the connection created between user and the other user as friend*/
+    public Friend(Integer friendID, Integer userID) {
         this.friendID = friendID;
-        this.friendName = friendName;
-        this.phoneNumber = phoneNumber;
-        this.profilePic = profilePic;
+        this.accountID = userID;
         this.blockedStatus = false;
-        this.userID = userID;
+
+        //TODO fetch the friend info from account based on their id then define the variable
+        this.friendName = null;
+        this.phoneNumber = null;
+        this.profilePic = null;
+    }
+
+    /*TODO insert connection between user to database as friend based on this.variable*/
+    public void addFriend(){
+
+    }
+
+    /*TODO remove connection between user to database as friend based on this.variable*/
+    public void removeFriend(){
+        String query = "DELETE FROM friends WHERE friend_id = [friend_id] AND user_id = [user_id]";
+    }
+
+    /*Fetch the friend information from database based on the accountID*/
+    public static List<Friend> fetchFriends(Integer accountID){
+        return null;
+    }
+
+    /*Friend blocked status*/
+    public void setBlockedStatus(boolean blockedStatus) {
+        this.blockedStatus = blockedStatus;
     }
 
     public boolean isBlockedStatus() {
         return blockedStatus;
     }
 
-    public void setBlockedStatus(boolean blockedStatus) {
-        this.blockedStatus = blockedStatus;
-    }
-
     public String getFriendName() {
         return friendName;
-    }
-
-    public void setFriendName(String friendName) {
-        this.friendName = friendName;
     }
 
     public String getPhoneNumber() {
         return phoneNumber;
     }
 
-    public void setPhoneNumber(String phoneNumber) {
-        this.phoneNumber = phoneNumber;
-    }
-
     public Blob getProfilePic() {
         return profilePic;
-    }
-
-    public void setProfilePic(Blob profilePic) {
-        this.profilePic = profilePic;
-    }
-
-    // TODO add remind function
-    // should the notification id be fixed
-    // get the friend debt
-    // if this friend is in debtlist this function  will create notify if not toast message
-    // "friendName is not in debt with you"
-
-    // TODO add settle up function
-    public void settleUpFriendDebt(Blob paymentEvidence){
-        SettleUpDebt settleUpDebt = new SettleUpDebt(friendID, userID, paymentEvidence);
-
-    }
-
-
-    // TODO get debt info
-    public Integer getDebtOwned(){
-        // TODO query to get friend info if the user is in debt
-        return 0;
     }
 
 }
