+++ app/src/main/java/com/example/coin_panion/classes/general/Account.java
@@ -1,73 +1,17 @@
 package com.example.coin_panion.classes.general;
 
-import com.example.coin_panion.classes.DebtLimit;
-import com.example.coin_panion.classes.Friend;
-import com.example.coin_panion.classes.SettleUp;
-import com.example.coin_panion.classes.User;
+import com.example.coin_panion.classes.notification.Notification;
+import com.example.coin_panion.classes.transaction.SettleUpAccount;
 
-import java.sql.Blob;
-import java.util.ArrayList;
 import java.util.List;
 
 public class Account {
+    private User user;
+    private String password;
+    private List<User> friends;
+    private DebtLimit debtLimit;
+    private SettleUpAccount settleUpAccount;
+    private List<Notification> notifications;
 
-//    private static User profile;
-//    private DebtLimit userDebtLimit;
-//    private List<PaymentRequest> paymentRequestsList;
-//    private List<Friend> friendsList;
-//    private List<SettleUp> paymentMethod;
-//
-//    public Account() {
-//        paymentRequestsList = new ArrayList<>();
-//        friendsList = new ArrayList<>();
-//        paymentMethod = new ArrayList<>();
-//        // Initialize the profile and userDebtLimit objects
-////        profile = new User();
-////        userDebtLimit = new DebtLimit();
-//    }
-//
-//    public void updateUserProfile(Blob profilePic, String userName, String phoneNumber, String email){
-//        if(profilePic != null) {
-//            profile.updateProfilePic(profilePic);
-//        }
-//        if(userName != null){
-//            profile.updateUserName(userName);
-//        }
-//        if(phoneNumber != null){
-//            profile.updatePhoneNumber(phoneNumber);
-//        }
-//        if(email != null){
-//            profile.updateEmail(email);
-//        }
-//    }
-//
-//    // TODO get the userID and DebtLimitID to perform update of debt limit
-//    public void setUserDebtLimit(String currencyType, String debtLimitTime, Integer debtLimitAmount){
-//        if(currencyType != null){
-//            userDebtLimit.setCurrencyType(currencyType);
-//        }
-//        if(debtLimitTime != null){
-//            userDebtLimit.setDebtLimitTime(debtLimitTime);
-//        }
-//        if(debtLimitAmount != null){
-//            userDebtLimit.setDebtLimitAmount(debtLimitAmount);
-//        }
-//    }
-//
-//    public List<Friend> getFriendsList() {
-//        return friendsList;
-//    }
-//
-//    public void setFriendsList(List<Friend> friendsList) {
-//        this.friendsList = friendsList;
-//    }
-//
-//    public List<SettleUp> getPaymentMethod() {
-//        return paymentMethod;
-//    }
-//
-//    public void setPaymentMethod(List<SettleUp> paymentMethod) {
-//        this.paymentMethod = paymentMethod;
-//    }
-
+    // Method to update user info
 }
