+++ app/src/main/java/com/example/coin_panion/classes/general/History.java
@@ -1,9 +1,7 @@
 package com.example.coin_panion.classes.general;
 
 import android.app.Notification;
-import com.example.coin_panion.classes.PaymentRequest;
-
-import com.example.coin_panion.classes.User;
+import com.example.coin_panion.classes.transaction.PaymentApproval;
 
 import java.util.ArrayList;
 
@@ -12,11 +10,11 @@ public class History {
     private String HistoryID;
     private Integer userID;
     ArrayList<Notification> notificationList = new ArrayList<>();
-    ArrayList<PaymentRequest> paymentRequestList = new ArrayList<>();
+    ArrayList<PaymentApproval> paymentApprovalList = new ArrayList<>();
 
-    public History(ArrayList<Notification> notificationList, ArrayList<PaymentRequest> paymentRequestList) {
+    public History(ArrayList<Notification> notificationList, ArrayList<PaymentApproval> paymentApprovalList) {
         this.notificationList = notificationList;
-        this.paymentRequestList = paymentRequestList;
+        this.paymentApprovalList = paymentApprovalList;
 //        this.userID = User.userID;
     }
 
@@ -26,7 +24,7 @@ public class History {
     }
 
     //TODO call Payment request from database based on the CreditorID that is the UserID place everthing into an arraylist
-    public ArrayList<PaymentRequest> DisplayPaymentRequestList(Integer userID){
+    public ArrayList<PaymentApproval> DisplayPaymentRequestList(Integer userID){
         return new ArrayList<>();
     }
