+++ app/src/main/java/com/example/coin_panion/classes/general/User.java
@@ -1,4 +1,4 @@
-package com.example.coin_panion.classes;
+package com.example.coin_panion.classes.general;
 
 import com.example.coin_panion.classes.utility.Line;
 import com.example.coin_panion.classes.utility.ThreadStatic;
