--- app/src/main/java/com/example/coin_panion/classes/group/Notification.java
@@ -1,4 +1,0 @@
-package com.example.coin_panion.classes.group;
-
-public class Notification{
-}
