+++ app/src/main/java/com/example/coin_panion/classes/transaction/SettleUpAccount.java
@@ -1,15 +1,15 @@
-package com.example.coin_panion.classes;
+package com.example.coin_panion.classes.transaction;
 
 import java.sql.Blob;
 
-public class SettleUp {
+public class SettleUpAccount {
     private Integer accountId;
     private String accountName;
     private long accountNumber;
     private Blob qrCode;
 
     /*Create new settle Up object for a user*/
-    public SettleUp(Integer accountId, String accountName, long accountNumber, Blob qrCode) {
+    public SettleUpAccount(Integer accountId, String accountName, long accountNumber, Blob qrCode) {
         this.accountId = accountId;
         this.accountName = accountName;
         this.accountNumber = accountNumber;
@@ -18,7 +18,7 @@ public class SettleUp {
 
 
     /*TODO fetch settleUp info from database based on accountID*/
-    public SettleUp fetchSettleUp(Integer accountId){
+    public SettleUpAccount fetchSettleUp(Integer accountId){
 
         return null;// fetch settleUp based info based on user accountId
     }
