+++ app/src/main/java/com/example/coin_panion/classes/transaction/Transaction.java
@@ -116,6 +116,8 @@ public class Transaction {
         return  listAtomicReference.get();
     };
 
+    // TODO settle up debt
+
 
     // Getters and Setters
     public Integer getTransactionID() {
