+++ app/src/main/java/com/example/coin_panion/classes/utility/BaseViewModel.java
@@ -3,10 +3,11 @@ package com.example.coin_panion.classes.utility;
 import androidx.lifecycle.ViewModel;
 
 import java.util.HashMap;
+import java.util.List;
 
 public class BaseViewModel extends ViewModel {
     private HashMap<String, Object> data = new HashMap<>();
-    private HashMap<String, Object[]> arrayData = new HashMap<>();
+    private HashMap<String, List<Object>> arrayData = new HashMap<>();
 
     public void put(String key, Object value){
         this.data.put(key, value);
@@ -16,11 +17,11 @@ public class BaseViewModel extends ViewModel {
         return this.data.get(key);
     }
 
-    public void putArray(String key, Object[] values){
+    public void putArray(String key, List<Object> values){
         this.arrayData.put(key, values);
     }
 
-    public Object[] getArray(String key){
+    public List<Object> getArray(String key){
         return this.arrayData.get(key);
     }
 }
